Implement User search based on name
This commit enable search of the User resource base on the first_name,
last_name and username attributes. We use a fuzzy search so that clients
are not required to know before hand exactly what to search for, they
can just use search term approximations. This approach is inspired by
how Google and the Github API implement search functionality.