Enable CORS
Our API server is being built to support requests from browser clients.
The browser clients have no requirement that they should run on the same
origin server as the API. Because of this, we need to enable
Cross-Origin Resource Sharing (CORS) so that the AJAX requests from the
browser clients, for resources on our server, may be successful.